Version: 2011-06-30 ksh93u+
Sourced from ast-open-history version 2011-06-30 and:
d7e0d892ec28195b7713c1ff8226f37cb23a65c32ed6e0835090682cad2ea6c1  INIT.2011-06-30.tgz

+++ src/cmd/INIT/ratz.c
@@ -5091,7 +5091,7 @@ char**	argv;
 				}
 				else
 				{
-					n = strlen(s = path);
+					n = (int)strlen(s = path);
 					p = (state.blocks * 100) / total;
 					if (n > (METER_width - METER_parts - 1))
 					{
